pyproject.toml
[project]
name = "Agentic-Multi-Domain-RAG-System"
version = "0.1.0"
description = "Agentic multi-domain RAG tutor bot with FastAPI, Streamlit, Qdrant, Redis, reranking, and web fallback."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "streamlit==1.38.0",
  "qdrant-client==1.11.3",
  "sentence-transformers==3.1.1",
  "rank-bm25==0.2.2",
  "PyMuPDF==1.24.10",
  "duckduckgo-search==6.2.11",
  "redis==5.0.8",
  "pydantic==2.9.2",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "python-multipart==0.0.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "app", "config", "services", "evaluation"]

[tool.setuptools.package-dir]
"" = "."